% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{alpha_calculation}
\alias{alpha_calculation}
\title{Calculate the matrix (alpha_{k,g})}
\usage{
alpha_calculation(s_ik = s_ik, t_jg = t_jg, nu_j = nu_j, mu_i = mu_i,
  K = K, G = G, x = x, exp_utilde = exp_utilde)
}
\arguments{
\item{s_ik}{s_ik.}

\item{t_jg}{t_jg.}

\item{nu_j}{nu_j.}

\item{mu_i}{mu_i.}

\item{K}{K.}

\item{G}{G.}

\item{x}{a matrix of observations. Columns correspond to biological samples and rows to microorganisms.}

\item{exp_utilde}{exp_utilde.}
}
\value{
a matrix of dimension (\code{K},\code{G}) of the terms of interactions.
}
\description{
Calculate the matrix (alpha_{k,g})
}
\keyword{internal}
