% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pam.R
\name{init_pam}
\alias{init_pam}
\title{Initialisation of the co-clusters by partitioning around medoids method.}
\usage{
init_pam(x, nu_j = NULL, a = NULL, K = K, G = G, akg = FALSE)
}
\arguments{
\item{x}{The output of the cobiclust function.}

\item{nu_j}{a vector of . The length is equal to the number of columns.}

\item{a}{an numeric.}

\item{K}{an integer specifying the number of groups in rows.}

\item{G}{an integer specifying the number of groups in columns.}

\item{akg}{a logical variable indicating whether to use a common dispersion parameter (akg = FALSE) or a dispersion parameter per cocluster (akg = TRUE).}
}
\value{
A list of
\describe{
\item{\code{nu_j}}{nu_j.}
\item{\code{mu_i}}{mu_i.}
\item{\code{t_jg}}{t_jg.}
\item{\code{s_ik}}{s_ik.}
\item{\code{pi_c}}{pi.}
\item{\code{rho_c}}{rho.}
\item{\code{a}}{a.}
\item{\code{exp_utilde}}{exp_utilde.}
\item{\code{exp_logutilde}}{exp_logutilde.}
\item{\code{alpha_c}}{alpha.}
}
}
\description{
Initialisation of the co-clusters by partitioning around medoids method.
}
\keyword{internal}
