% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{lb_calculation}
\alias{lb_calculation}
\title{Calculate the lower bound}
\usage{
lb_calculation(x = x, qu_param = qu_param, s_ik = s_ik, pi_c = pi_c,
  t_jg = t_jg, rho_c = rho_c, mu_i = mu_i, nu_j = nu_j,
  alpha_c = alpha_c, a = a, akg = TRUE)
}
\arguments{
\item{x}{a matrix of observations. Columns correspond to biological samples and rows to microorganisms.}

\item{s_ik}{s_ik.}

\item{pi_c}{pi_c.}

\item{t_jg}{t_jg.}

\item{rho_c}{rho_c.}

\item{mu_i}{mu_i.}

\item{nu_j}{nu_j.}

\item{alpha_c}{a matrix the terms of interactions.}

\item{a}{a.}

\item{akg}{a logical variable indicating whether to use a common dispersion parameter (akg = FALSE) or a dispersion parameter per cocluster (akg = TRUE).}

\item{qu.param}{qu.param.}
}
\value{
a list of 2 elements.
\describe{
\item{\code{lb}}{value of the lower bound.}
\item{\code{ent}}{value of the entropy term.}
}
}
\description{
Calculate the lower bound
}
\keyword{internal}
