% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_criteria.R
\name{selection_criteria}
\alias{selection_criteria}
\title{Calculate selection criteria.}
\usage{
selection_criteria(x, K, G)
}
\arguments{
\item{x}{The output of the cobiclust function.}

\item{K}{The number of groups in rows.}

\item{G}{The number of groups in columns.}
}
\value{
A dataframe with 7 columns.
\describe{
\item{\code{vICL}}{the vICL selection criterion.}
\item{\code{BIC}}{the BIC selection criterion.}
\item{\code{penKG}}{the value of the BIC penalty.}
\item{\code{lb}}{the value of the lower bound of the log-likelihood.}
\item{\code{entZW}}{the value of the entropy of the latent variables Z and W.}
\item{\code{K}}{the number of groups in rows.}
\item{\code{G}}{the number of groups in columns.}
}
}
\description{
Calculate selection criteria.
}
