% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobiclustering.R
\name{summary.cobiclustering}
\alias{summary.cobiclustering}
\title{Summary of an object of class Cobiclust}
\usage{
\method{summary}{cobiclustering}(object, ...)
}
\arguments{
\item{object}{an object of class cobiclustering.}

\item{...}{ignored}
}
\description{
Summary of an object of class Cobiclust
}
\keyword{internal}
